requirements.txt
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage==0.36.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
